refactor: relocate and clean up core DI annotations and qualifiers
