+++ core/base/src/commonMain/kotlin/com/thomaskioko/tvmaniac/core/base/ActivityScope.kt
@@ -1,0 +1,3 @@
+package com.thomaskioko.tvmaniac.core.base
+
+public abstract class ActivityScope private constructor()
