+++ core/base/src/commonMain/kotlin/com/thomaskioko/tvmaniac/core/base/Qualifiers.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.tvmaniac.core.base.di
+package com.thomaskioko.tvmaniac.core.base
 
 import dev.zacsweers.metro.Qualifier
 import kotlin.annotation.AnnotationRetention.BINARY
