--- core/base/src/commonMain/kotlin/com/thomaskioko/tvmaniac/core/base/annotations/ActivityScope.kt
@@ -1,6 +1,0 @@
-package com.thomaskioko.tvmaniac.core.base.annotations
-
-import me.tatarka.inject.annotations.Scope
-
-@Scope
-public annotation class ActivityScope
