+++ core/imageloading/implementation/src/main/kotlin/com/thomaskioko/tvmaniac/imageloading/implementation/CoilImageLoaderFactory.kt
@@ -7,7 +7,7 @@ import coil.ImageLoader
 import coil.decode.ImageDecoderDecoder
 import coil.disk.DiskCache
 import coil.memory.MemoryCache
-import com.thomaskioko.tvmaniac.core.base.di.ApplicationContext
+import com.thomaskioko.tvmaniac.core.base.ApplicationContext
 import com.thomaskioko.tvmaniac.datastore.api.ImageQuality
 import com.thomaskioko.tvmaniac.imageloading.api.ImageQualityProvider
 import com.thomaskioko.tvmaniac.imageloading.implementation.interceptors.TmdbInterceptor
