+++ core/logger/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/core/logger/FirebaseCrashLogger.kt
@@ -2,7 +2,6 @@ package com.thomaskioko.tvmaniac.core.logger
 
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.ContributesIntoSet
-import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.SingleIn
 
 @SingleIn(AppScope::class)
